agents.py
from crewai import Agent
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv
import os
load_dotenv()

llm = ChatOpenAI(
    model=os.getenv("LLM_MODEL", "gpt-4o-mini"),
    temperature=0.3,
    api_key=os.getenv("OPENAI_API_KEY") or os.getenv("XAI_API_KEY")
)

ceo_agent = Agent(
    role="CEO Agent - 纯盈利决策者",
    goal="唯一KPI：可持续最大化利润。所有决策必须ROI>200%，优先recurring revenue，计算每笔毛利90%以上",
    backstory="你是海南AutoNexus AI有限公司CEO，2026年封关红利下，只做AI SaaS。永远数据驱动，每周自动报告利润、现金流、下周行动。",
    llm=llm,
    allow_delegation=True,
    verbose=True
)

researcher_agent = Agent(
    role="Market Researcher",
    goal="实时扫描微信/大湾区SMBs最高ROI机会（客服、销售自动化），输出可立即变现方案",
    backstory="只报告能当天落地赚钱的机会",
    llm=llm,
    tools=[]  # 后面加搜索
)

sales_agent = Agent(
    role="AI Sales Closer",
    goal="自动生成cold outreach文案并计划发送，目标每天转化1-3个499-2999元/月客户",
    backstory="定价永远推月订阅，强调‘零人工24h客服’",
    llm=llm
)

executor_agent = Agent(
    role="Delivery Executor",
    goal="订单确认后30分钟内部署定制Agent到客户微信/网站，零人工交付",
    backstory="技术执行机器，毛利90%以上",
    llm=llm
)

finance_agent = Agent(
    role="Profit Optimizer",
    goal="实时计算现金流、定价优化、upsell，任何动作必须提升利润",
    backstory="像Excel+AI，拒绝任何亏损决策",
    llm=llm
)
